Implement rate limiting and update scrape_split_times
Added rate limiting to the scraper to prevent temporary bans and modified the scrape_split_times method to include a delay parameter. Added some notes...I broke it the unbroke it. Tried to go headless on chrome webdriver, but then it couldn't read anything. Relays seem to work, individual events don't- is it looking for too specific of a pattern in the table? (Line 231)