Found and fixed issue with frames in selenium
Initially couldn't find any of the .htm links. Selenium needs to be directed to which frame to search when trying to find_elements, as it treats each frame as a separate document (or something of the like). So searched for the frame, swapped to it, now can find all the links appropriately. Swapped to original frame back after function ends, to be safe.

Also added category for relay-only sessions.

Also changed logic for finding which line to start looking for results, as it turns out the "=======" always happens twice, but can change length depending on event. So now instead of a direct match it uses if line.startswith('======')

Update version number. Among other things ;)